Added: Interface and Adapters

+++ database/migrations/2020_02_14_200503_create_currency_records_table.php
@@ -15,8 +15,8 @@ class CreateCurrencyRecordsTable extends Migration
     {
         Schema::create('currency_records', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigInteger('currency_id')->unsigned();
-            $table->bigInteger('amount');
+            $table->integer('currency_id')->unsigned();
+            $table->float('amount', 8, 5);
             $table->timestamps();
 
             $table->foreign('currency_id')->references('id')->on('currencies');
